+++ service/tbc/indexer.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2025 Hemi Labs, Inc.
+// Copyright (c) 2025-2026 Hemi Labs, Inc.
 // Use of this source code is governed by the MIT License,
 // which can be found in the LICENSE file.
 
@@ -84,6 +84,7 @@ type indexer interface {
 	process(ctx context.Context, direction int, block *btcutil.Block, c indexerCache) error // Process block
 	commit(ctx context.Context, direction int, hash chainhash.Hash, c indexerCache) error   // Commit index cache to disk
 	fixupCacheHook(ctx context.Context, block *btcutil.Block, c indexerCache) error         // Fixup cache
+	onSyncComplete()                                                                        // Called after sync reaches target
 }
 
 // indexerCache exposes Cache management functions.
@@ -129,7 +130,11 @@ func (c *indexerCommon) IndexToBest(ctx context.Context) error {
 	}
 	defer c.indexing.Store(0)
 
-	return c.toBest(ctx)
+	if err := c.toBest(ctx); err != nil {
+		return err
+	}
+	c.p.onSyncComplete()
+	return nil
 }
 
 func (c *indexerCommon) IndexToHash(ctx context.Context, hash chainhash.Hash) error {
@@ -143,7 +148,11 @@ func (c *indexerCommon) IndexToHash(ctx context.Context, hash chainhash.Hash) er
 	}
 	defer c.indexing.Store(0)
 
-	return c.windOrUnwind(ctx, hash)
+	if err := c.windOrUnwind(ctx, hash); err != nil {
+		return err
+	}
+	c.p.onSyncComplete()
+	return nil
 }
 
 func (c *indexerCommon) IndexerAt(ctx context.Context) (*tbcd.BlockHeader, error) {